+++ src/Database/Design/Ampersand/Misc/Options.hs
@@ -1,4 +1,5 @@
-{-# LANGUAGE PatternGuards #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE OverloadedStrings #-}
 module Database.Design.Ampersand.Misc.Options
         (Options(..),getOptions,usageInfo'
         ,verboseLn,verbose,FSpecFormat(..)
@@ -15,10 +16,13 @@ import Data.Time.LocalTime
 import Control.Monad
 import Data.Maybe
 import Database.Design.Ampersand.Basics
-import Paths_ampersand (getDataDir)
 import Prelude hiding (writeFile,readFile,getContents,putStr,putStrLn)
 import Data.List
 import Data.Char
+
+import Data.Yaml
+import GHC.Generics
+import Data.Either
 
 -- | This data constructor is able to hold all kind of information that is useful to
 --   express what the user would like Ampersand to do.
@@ -31,6 +35,7 @@ data Options = Options { showVersion :: Bool
                        , validateSQL :: Bool
                        , genPrototype :: Bool
                        , dirPrototype :: String  -- the directory to generate the prototype in.
+                       , dirInclude :: String -- the directory that is included in the generated prototype
                        , allInterfaces :: Bool
                        , dbName :: String
                        , namespace :: String
@@ -61,7 +66,6 @@ data Options = Options { showVersion :: Bool
                        , genBericht :: Bool
                        , language :: Maybe Lang  -- The language in which the user wants the documentation to be printed.
                        , dirExec :: String --the base for relative paths to input files
-                       , ampersandDataDir :: FilePath -- the directory where Ampersand data files are.
                        , progrName :: String --The name of the adl executable
                        , fileName :: FilePath --the file with the Ampersand context
                        , baseName :: String
@@ -84,40 +88,31 @@ getOptions =
    do args     <- getArgs
       progName <- getProgName
       exePath  <- getExecutablePath -- findExecutable progName
-      dataPath <- getDataDir
-      let haskellInstallationDirectory = dataPath </> "AmpersandData"
-      existshaskellInstallationDirectory <- doesDirectoryExist  haskellInstallationDirectory
---      putStrLn  $ "haskellInstallationDirectory: "  ++haskellInstallationDirectory
-
-      let ampersandInstallationDirectory = takeDirectory exePath </> ".." </> "AmpersandData"
-      existsampersandInstallationDirectory <- doesDirectoryExist  ampersandInstallationDirectory
---      putStrLn  $ "ampersandInstallationDirectory: "++ampersandInstallationDirectory
-
-      let dataDir
-            | existshaskellInstallationDirectory   = haskellInstallationDirectory
-            | existsampersandInstallationDirectory = ampersandInstallationDirectory
-            | otherwise                            = exePath
       localTime <-  do utcTime <- getCurrentTime
                        timeZone <- getCurrentTimeZone
                        return (utcToLocalTime timeZone utcTime)
       env <- getEnvironment
       let usage = "\nType '"++ progName++" --help' for usage info."
-      let (actions, fNames, errors) = getOpt Permute (map fst options) args
-      when ((not.null) errors) (error $ concat errors ++ usage)
+          (configs,plainArgs) = partition (isPrefixOf "--config") args
+      argsFromYaml <- mapM readYamlConfig configs
+      let (actions, fNames, errors) = getOpt Permute (map fst options) $ concat argsFromYaml ++ plainArgs
+      
+      unless (null errors) (error $ concat errors ++ usage)
       let fName = case fNames of
                    []  -> error ("Please supply the name of an ampersand file" ++ usage)
                    [x] -> x
                    _   -> error ("too many files: "++ intercalate ", " fNames  ++ usage)
 
-      let startOptions =
+      let startOptions :: Options
+          startOptions =
                Options {genTime          = localTime
                       , dirOutput        = fromMaybe "."       (lookup envdirOutput    env)
                       , outputfile       = fatal 83 "No monadic options available."
-                      , dirPrototype     = fromMaybe ("." </> (addExtension (takeBaseName fName) ".proto"))
-                                                     (lookup envdirPrototype env) </> (addExtension (takeBaseName fName) ".proto")
+                      , dirPrototype     = fromMaybe ("." </> addExtension (takeBaseName fName) ".proto")
+                                                     (lookup envdirPrototype env) </> addExtension (takeBaseName fName) ".proto"
+                      , dirInclude       = "include"
                       , dbName           = map toLower $ fromMaybe ("ampersand_"++takeBaseName fName) (lookup envdbName env)
                       , dirExec          = takeDirectory exePath
-                      , ampersandDataDir = dataDir
                       , preVersion       = fromMaybe ""        (lookup "CCPreVersion"  env)
                       , postVersion      = fromMaybe ""        (lookup "CCPostVersion" env)
                       , showVersion      = False
@@ -169,6 +164,8 @@ getOptions =
                       , trimXLSXCells    = True
                       }
       -- Here we thread startOptions through all supplied option actions
+      
+      
       opts <- foldl (>>=) (return startOptions) actions
       -- Now we do some checks on the options:
       when (development opts && validateSQL opts)
@@ -177,7 +174,64 @@ getOptions =
       when (genPrototype opts)
            (createDirectoryIfMissing True (dirPrototype opts))
       return opts
+  where
+      readYamlConfig :: String -> IO [String]
+      readYamlConfig ('-':'-':'c':'o':'n':'f':'i':'g':'=':yaml)  
+         = do putStrLn $ "config file:"++yaml
+              res <- parseYaml yaml
+              case res of 
+                  Left err -> do putStrLn $ "." </> yaml ++" could not be parsed."
+                                 putStrLn $ prettyPrintParseException err
+                                 return $ error "Please fix this first."
+                  Right (Config args)
+                           -> do print args
+                                 case partitionEithers . map arg2str $ args of
+                                   (xs ,[]) -> return xs
+                                   (_, x:_) -> error $ "Error in config file `"++yaml++"`: "++x
+        where
+          arg2str :: Argument -> Either String String
+          arg2str (Argument str) 
+            | "config" `isPrefixOf` str = Right "`config` is not allowed inside a config file"
+            | otherwise = Left $ "--"++str
+      readYamlConfig "--config"
+         = error "Command line argument --config requires an file name as an argument"
+      readYamlConfig str
+         = error $ "Cannot handle command line argument: "++str 
 
+      parseYaml :: String -> IO (Either ParseException Config) 
+      parseYaml yaml = decodeFileEither $ "." </> yaml
+data Config = Config [Argument] deriving (Generic,Show)
+instance FromJSON Config
+data Argument = Argument String deriving (Generic,Show)
+instance FromJSON Argument
+
+writeConfigFile :: IO ()
+writeConfigFile = writeFile sampleConfigFileName (unlines sampleConfigFile)
+sampleConfigFileName :: FilePath
+sampleConfigFileName = "sampleconfig.yaml"
+sampleConfigFile :: [String]
+sampleConfigFile =
+  [ " # Sample config file for Ampersand"
+  , " # This file contains a list of all command line options that can be set using a config file"
+  , " # It can be used by specifying:  ampersand.exe --config=myConfig.yaml myModel.adl"
+  , " # remove the comment character (`#`) in front of the switches you want to activate."
+  , " # Note: Make sure that the minus (`-`) characters are in exactly the same column. Yaml format is picky about that."
+  , ""
+  , "config:"
+  ]++(concatMap yamlItem . filter canBeYamlOption . map fst $ options)
+  where
+    yamlItem :: OptionDef -> [String]
+    yamlItem (Option _ label kind info ) =
+      [ " ### "++info++":"
+      , " # - "++head label++case kind of
+                                 NoArg _ -> "" 
+                                 ReqArg _ str -> "="++str
+                                 OptArg _ str -> "[="++str++"]"
+      , ""
+      ]
+    canBeYamlOption :: OptionDef -> Bool
+    canBeYamlOption (Option _ label _ _ ) =
+      null $ label `isc` ["version","help","config","sampleConfigFile"]  
 data DisplayMode = Public | Hidden deriving Eq
 
 data FSpecFormat = FPandoc| Fasciidoc| Fcontext| Fdocbook| Fhtml| FLatex| Fman| Fmarkdown| Fmediawiki| Fopendocument| Forg| Fplain| Frst| Frtf| Ftexinfo| Ftextile deriving (Show, Eq)
@@ -205,6 +259,17 @@ options = [ (Option ['v']   ["version"]
                (NoArg (\opts -> return opts{verboseP = True}))
                "verbose error message format."
             , Public)
+          , (Option []   ["sampleConfigFile"]
+               (NoArg (\opts -> do writeConfigFile
+                                   putStrLn (sampleConfigFileName++" written.")
+                                   return opts))
+               ("write a sample configuration file ("++sampleConfigFileName++")")
+            , Public)
+          , (Option []      ["config"]
+               (ReqArg (\nm _ -> fatal 194 $ "config file ("++nm++")should not be treated as a regular option."
+                       ) "config.yaml")
+               "config file (*.yaml)"
+            , Public)
           , (Option []      ["dev"]
                (NoArg (\opts -> return opts{development = True}))
                "Report and generate extra development information (for Martijn)"
@@ -219,6 +284,12 @@ options = [ (Option ['v']   ["version"]
                        ) "DIRECTORY")
                ("generate a functional prototype (overrules environment variable "++ envdirPrototype ++ ").")
             , Public)
+          , (Option []     ["include"]
+               (ReqArg (\nm opts -> return opts {dirInclude = nm
+                                                ,genPrototype = True}
+                       ) "DIRECTORY")
+               "include a directory into the generated prototype, instead of the default."
+            , Public)
           , (Option ['d']  ["dbName"]
                (ReqArg (\nm opts -> return opts{dbName = if nm == ""
                                                          then dbName opts
@@ -231,19 +302,19 @@ options = [ (Option ['v']   ["version"]
                                                           then sqlHost opts
                                                           else nm}
                        ) "HOSTNAME")
-               ("set SQL host name (Defaults to `localhost`)")
+               "set SQL host name (Defaults to `localhost`)"
             , Public)
           , (Option []  ["sqlLogin"]
                (ReqArg (\nm opts -> return opts{sqlLogin = if nm == ""
                                                           then sqlLogin opts
                                                           else nm}
                        ) "USER")
-               ("set SQL host name (Defaults to `ampersand`)")
+               "set SQL user name (Defaults to `ampersand`)"
             , Public)
           , (Option []  ["sqlPwd"]
                (ReqArg (\nm opts -> return opts{sqlPwd = nm}
                        ) "PASSWORD")
-               ("set SQL host name (Defaults to `ampersand`)")
+               "set SQL password (Defaults to `ampersand`)"
             , Public)
           , (Option []        ["sql-bin-tables"]
                (NoArg (\opts -> return opts{sqlBinTables = True}))
@@ -428,7 +499,7 @@ infoHeader progName = "\nUsage info:\n " ++ progName ++ " options file ...\n\nLi
 publishOption:: OptDescr a -> String
 publishOption (Option shorts longs args expl) 
   = unlines (
-    [ "  "++intercalate ", " ["--"++l | l <-longs] 
+    ( "  "++intercalate ", " ["--"++l | l <-longs] 
       ++case args of
          NoArg _      -> "" 
          ReqArg _ str -> "="++str
@@ -436,8 +507,8 @@ publishOption (Option shorts longs args expl)
       ++case intercalate ", " [ "-"++[c] | c <- shorts] of
           []  -> []
           xs  -> " ("++xs++")"
-    ] 
-    ) ++unlines (map (replicate 10 ' '++) (lines (limit 65 expl)))
+    ): 
+     map (replicate 10 ' '++) (lines (limit 65 expl)))
   where
    limit :: Int -> String -> String
    limit i = intercalate "\n" . map (singleLine i . words) . lines